TM_Server uses NC_Server backend the right way
TM_Server was re-implementing some of the functionality already
implemented in the NC_Server class. The portions re-implemented were
related to the ability to organize and use multiple clients. TM_Server
now uses interfaces and the provided organization inside of NC_Server for storing
multiple client sockets and using them.

However, a vector of clients is also stored in TM_Server to track
TM-related aspects of each client. Therefore, the partitioning of the
class functionality is that TM_Server deals with mitigating conflicts
and other TM related functionality, while the NC_Server deals with the
server-client communication at the lower level.

This is a functionality correction fix, a maintenance commit is required
to clean up some of the legacy functionality and commented code.