Corrected SendMessage network send; Client struct index
Refactored code to use the index of the connected client that the thread
is dealing with. This was previously done with a reference to the
element. The pointer method was a bit sloppy and assumed that the vector
would not be moving any elements around in memory.

The SendMessage method was not passing a client id to its send call.
This overloaded method in NC_Server just uses the last client's socket
and buffer to send data. This was causeing some major issues with
concurrent clients.

Additionally, in debugging the above problems, many new debug statements
were added. These need to be wrapped in #ifdef DEBUG statements.