Added use of new access cache; detect remote disp.
Server now uses the hardware analogous software access cache. Tested with
mutex operation and things seem to be working.

Added a check for a display client connecting. The display client will
receive updates from the threads on the operations of the node
transactions. This way we can graph and make pretty visuals.